requirements.txt
locust==1.2.2
Faker==4.1.2
sqlalchemy==1.3.19
slackclient==2.8.0
pymysql==0.10.0
